feature-fix

+++ pages/mart/omni7/index.js
@@ -249,6 +249,9 @@ function MartOmni7Page(props) {
    }
 
    const handleDeleteEditModal = async (id) => {
+      if (!window.confirm("คุณแน่ใจที่จะลบใช่หรือไม่")) {
+         return 
+      }
       try {
          const response = await fetch(`/api/mart/omni7?id=${id}`, {
             method: "DELETE",
