+++ pages/mart/promotion/index.js
@@ -258,6 +258,9 @@ function MartPromotionPage(props) {
    }
 
    const handleDeleteEditModal = async (id) => {
+      if (!window.confirm("คุณแน่ใจที่จะลบใช่หรือไม่")) {
+         return 
+      }
       try {
          const response = await fetch(`/api/mart/promotion?id=${id}`, {
             method: "DELETE",
