+++ pages/tracking/yahoo/index.js
@@ -139,6 +139,9 @@ function YahooTrackingsPage(props) {
       }
    }
    const handleDeleteRow = async (id) => {
+      if (!window.confirm("คุณแน่ใจที่จะลบใช่หรือไม่")) {
+         return 
+      }
       try {
          const response = await fetch(`/api/tracking/yahoo?id=${id}`, {
             method: "DELETE",
