+++ pages/yahoo/payment/index.js
@@ -39,6 +39,9 @@ function YahooPaymentPage() {
    const [showSlipModal, setShowSlipModal] = useState(false)
 
    const handleDeleteRow = async (id) => {
+      if (!window.confirm("คุณแน่ใจที่จะลบใช่หรือไม่")) {
+         return 
+      }
       try {
          const response = await fetch("/api/yahoo/payment", {
             method: "DELETE",
